#28 - UI: Move insight view into another tab
